requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cffi==1.14.5
cmdstanpy==0.9.68
convertdate==2.3.2
cycler==0.10.0
Cython==0.29.23
decorator==5.0.7
defusedxml==0.7.1
entrypoints==0.3
ephem==3.7.7.1
hijri-converter==2.1.1
holidays==0.11.1
importlib-metadata==4.0.1
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pmdarima==1.8.2
prometheus-client==0.10.1
prompt-toolkit==3.0.18
prophet==1.0.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
PyMeeus==0.5.11
pyparsing==2.4.7
pyrsistent==0.17.3
pystan==2.19.1.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
setuptools-git==1.2
six==1.15.0
sklearn==0.0
statsmodels==0.12.2
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.60.0
traitlets==4.3.3
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.1
